requirements.txt
flask
gevent-socketio
gevent
RPi.GPIO
